+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 All notable changes to the `ci-realredis` library will be documented in this change log file.
 
-## v1.1.8 - 2018-12-03
+## v1.1.9 - 2018-12-03
+
+### Added
+- getAndSet() method to `SupportRedis`
+
+### Modified
+- Removed error_log statements from the getElseSet() method
+
+## v1.1.8 - 2018-12-02
 
 ### Added
 - Post install composer script